TeleService: Do not create a PhoneAccount for the slot without SIM card
When the slot has no SIM card, the dummy subId now is no longer
negative number, so the TelecomAccountRegistry will create a PhoneAccount
for the slot without SIM card too.
Check whether the slot has SIM card and it is provisoned. if yes, add a
PhoneAccount for the slot with SIM card.

Change-Id: I19934f735ab0d77a9f464ebb90d11574c8c16955
CRs-Fixed: 820101
Signed-off-by: sub77 <sub77@ymail.com>

Conflicts:
	src/com/android/services/telephony/TelecomAccountRegistry.java